fix: 시간관련 커스텀 훅으로 분리
